Import only translations for currently enabled features
There are a lot of translations in total, so it could be overwhelming for
the user to figure out which translations are actually being used for
the current set of enabled features.

So, we import only translations for features that are currently enabled,
and allow re-importing while keeping existing translations intact.